test: split oversized unit fixtures
