+++ tests/SemanticStub.Api.Tests/Unit/Inspection/StubInspectionServiceTests.cs
@@ -252,266 +252,9 @@ public sealed class StubInspectionServiceTests
     // GetRuntimeMetrics
     // ---------------------------------------------------------------------------
 
-    [Fact]
-    public void GetRuntimeMetrics_ReturnsZeroValues_BeforeRequestsAreRecorded()
-    {
-        var service = CreateService(EmptyDocument());
-
-        var metrics = service.GetRuntimeMetrics();
-
-        Assert.Equal(0, metrics.TotalRequestCount);
-        Assert.Equal(0, metrics.MatchedRequestCount);
-        Assert.Equal(0, metrics.UnmatchedRequestCount);
-        Assert.Equal(0, metrics.FallbackResponseCount);
-        Assert.Equal(0, metrics.SemanticMatchCount);
-        Assert.Equal(0, metrics.AverageLatencyMilliseconds);
-        Assert.Empty(metrics.StatusCodes);
-        Assert.Empty(metrics.TopRoutes);
-    }
-
-    [Fact]
-    public void RecordRequestMetrics_UpdatesMatchedCounts_StatusCodes_AndRoutes()
-    {
-        var service = CreateService(EmptyDocument());
-
-        service.RecordRequestMetrics(
-            CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "listUsers", matchMode: "fallback"),
-            StatusCodes.Status200OK,
-            TimeSpan.FromMilliseconds(25));
-
-        var metrics = service.GetRuntimeMetrics();
-
-        Assert.Equal(1, metrics.TotalRequestCount);
-        Assert.Equal(1, metrics.MatchedRequestCount);
-        Assert.Equal(0, metrics.UnmatchedRequestCount);
-        Assert.Equal(1, metrics.FallbackResponseCount);
-        Assert.Equal(0, metrics.SemanticMatchCount);
-        Assert.Equal(25, metrics.AverageLatencyMilliseconds);
-        Assert.Collection(metrics.StatusCodes, entry =>
-        {
-            Assert.Equal(StatusCodes.Status200OK, entry.StatusCode);
-            Assert.Equal(1, entry.RequestCount);
-        });
-        Assert.Collection(metrics.TopRoutes, entry =>
-        {
-            Assert.Equal("listUsers", entry.RouteId);
-            Assert.Equal(1, entry.RequestCount);
-        });
-    }
-
-    [Fact]
-    public void RecordRequestMetrics_UpdatesUnmatchedAndSemanticCounts()
-    {
-        var service = CreateService(EmptyDocument());
-
-        service.RecordRequestMetrics(
-            CreateRecordedExplanation(matched: false, matchResult: "PathNotFound"),
-            StatusCodes.Status404NotFound,
-            TimeSpan.FromMilliseconds(10));
-        service.RecordRequestMetrics(
-            CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "searchUsers", matchMode: "semantic"),
-            StatusCodes.Status200OK,
-            TimeSpan.FromMilliseconds(30));
-
-        var metrics = service.GetRuntimeMetrics();
-
-        Assert.Equal(2, metrics.TotalRequestCount);
-        Assert.Equal(1, metrics.MatchedRequestCount);
-        Assert.Equal(1, metrics.UnmatchedRequestCount);
-        Assert.Equal(0, metrics.FallbackResponseCount);
-        Assert.Equal(1, metrics.SemanticMatchCount);
-        Assert.Equal(20, metrics.AverageLatencyMilliseconds);
-    }
-
-    [Fact]
-    public void GetRuntimeMetrics_OrdersStatusCodesAndRoutes_ByUsage()
-    {
-        var service = CreateService(EmptyDocument());
-
-        service.RecordRequestMetrics(
-            CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "b-route"),
-            StatusCodes.Status200OK,
-            TimeSpan.FromMilliseconds(10));
-        service.RecordRequestMetrics(
-            CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "a-route"),
-            StatusCodes.Status404NotFound,
-            TimeSpan.FromMilliseconds(20));
-        service.RecordRequestMetrics(
-            CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "a-route"),
-            StatusCodes.Status404NotFound,
-            TimeSpan.FromMilliseconds(30));
-
-        var metrics = service.GetRuntimeMetrics();
-
-        Assert.Collection(
-            metrics.StatusCodes,
-            first =>
-            {
-                Assert.Equal(StatusCodes.Status404NotFound, first.StatusCode);
-                Assert.Equal(2, first.RequestCount);
-            },
-            second =>
-            {
-                Assert.Equal(StatusCodes.Status200OK, second.StatusCode);
-                Assert.Equal(1, second.RequestCount);
-            });
-        Assert.Collection(
-            metrics.TopRoutes,
-            first =>
-            {
-                Assert.Equal("a-route", first.RouteId);
-                Assert.Equal(2, first.RequestCount);
-            },
-            second =>
-            {
-                Assert.Equal("b-route", second.RouteId);
-                Assert.Equal(1, second.RequestCount);
-            });
-    }
-
-    [Fact]
-    public void StubInspectionRuntimeStore_RecordRequestMetrics_ClampsNegativeLatency()
-    {
-        var store = new StubInspectionRuntimeStore();
-
-        store.RecordRequestMetrics(
-            CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "listUsers"),
-            StatusCodes.Status200OK,
-            TimeSpan.FromMilliseconds(-10));
-
-        var metrics = store.GetRuntimeMetrics();
-
-        Assert.Equal(0, metrics.AverageLatencyMilliseconds);
-        Assert.Collection(metrics.TopRoutes, route => Assert.Equal("listUsers", route.RouteId));
-    }
-
-    [Fact]
-    public void ResetRuntimeMetrics_ClearsMetricsAndRecentRequests_ButKeepsLastMatchExplanation()
-    {
-        var service = CreateService(EmptyDocument());
-        var explanation = CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "listUsers", matchMode: "fallback");
-
-        service.RecordRequestMetrics(explanation, StatusCodes.Status200OK, TimeSpan.FromMilliseconds(25));
-        service.RecordRecentRequest(
-            DateTimeOffset.Parse("2026-04-07T00:00:00Z"),
-            HttpMethods.Get,
-            "/users",
-            explanation,
-            StatusCodes.Status200OK,
-            TimeSpan.FromMilliseconds(25));
-        service.RecordLastMatchExplanation(explanation);
-
-        service.ResetRuntimeMetrics();
-
-        var metrics = service.GetRuntimeMetrics();
-        Assert.Equal(0, metrics.TotalRequestCount);
-        Assert.Equal(0, metrics.MatchedRequestCount);
-        Assert.Equal(0, metrics.UnmatchedRequestCount);
-        Assert.Equal(0, metrics.FallbackResponseCount);
-        Assert.Equal(0, metrics.SemanticMatchCount);
-        Assert.Equal(0, metrics.AverageLatencyMilliseconds);
-        Assert.Empty(metrics.StatusCodes);
-        Assert.Empty(metrics.TopRoutes);
-        Assert.Empty(service.GetRecentRequests(20));
-        Assert.Same(explanation, service.GetLastMatchExplanation());
-    }
-
     // ---------------------------------------------------------------------------
     // GetRecentRequests
     // ---------------------------------------------------------------------------
-
-    [Fact]
-    public void GetRecentRequests_ReturnsEmpty_WhenNothingHasBeenRecorded()
-    {
-        var service = CreateService(EmptyDocument());
-
-        var requests = service.GetRecentRequests(20);
-
-        Assert.Empty(requests);
-    }
-
-    [Fact]
-    public void RecordRecentRequest_ReturnsNewestFirst_AndCapturesFailureReason()
-    {
-        var service = CreateService(EmptyDocument());
-
-        service.RecordRecentRequest(
-            DateTimeOffset.Parse("2026-04-07T00:00:00Z"),
-            HttpMethods.Get,
-            "/users",
-            new MatchExplanationInfo
-            {
-                SelectionReason = "No configured route matched the supplied request path.",
-                Result = new MatchSimulationInfo
-                {
-                    Matched = false,
-                    MatchResult = "PathNotFound",
-                }
-            },
-            StatusCodes.Status404NotFound,
-            TimeSpan.FromMilliseconds(15));
-        service.RecordRecentRequest(
-            DateTimeOffset.Parse("2026-04-07T00:00:01Z"),
-            HttpMethods.Post,
-            "/orders",
-            CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "createOrder", matchMode: "fallback"),
-            StatusCodes.Status201Created,
-            TimeSpan.FromMilliseconds(25));
-
-        var requests = service.GetRecentRequests(20);
-
-        Assert.Equal(2, requests.Count);
-        Assert.Equal("/orders", requests[0].Path);
-        Assert.Equal("createOrder", requests[0].RouteId);
-        Assert.Equal("fallback", requests[0].MatchMode);
-        Assert.Null(requests[0].FailureReason);
-        Assert.Equal("/users", requests[1].Path);
-        Assert.Equal("No configured route matched the supplied request path.", requests[1].FailureReason);
-    }
-
-    [Fact]
-    public void GetRecentRequests_RespectsLimit_AndRetentionBound()
-    {
-        var service = CreateService(EmptyDocument());
-
-        for (var index = 0; index < 105; index++)
-        {
-            service.RecordRecentRequest(
-                DateTimeOffset.UtcNow.AddSeconds(index),
-                HttpMethods.Get,
-                $"/requests/{index}",
-                CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: $"route-{index}"),
-                StatusCodes.Status200OK,
-                TimeSpan.FromMilliseconds(index));
-        }
-
-        var limited = service.GetRecentRequests(3);
-        var all = service.GetRecentRequests(200);
-
-        Assert.Equal(3, limited.Count);
-        Assert.Equal("/requests/104", limited[0].Path);
-        Assert.Equal("/requests/102", limited[2].Path);
-        Assert.Equal(100, all.Count);
-        Assert.Equal("/requests/104", all[0].Path);
-        Assert.Equal("/requests/5", all[^1].Path);
-    }
-
-    [Fact]
-    public void StubInspectionRuntimeStore_GetRecentRequests_ReturnsEmpty_WhenLimitIsZeroOrNegative()
-    {
-        var store = new StubInspectionRuntimeStore();
-
-        store.RecordRecentRequest(
-            DateTimeOffset.Parse("2026-04-07T00:00:00Z"),
-            HttpMethods.Get,
-            "/users",
-            CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "listUsers"),
-            StatusCodes.Status200OK,
-            TimeSpan.FromMilliseconds(15));
-
-        Assert.Empty(store.GetRecentRequests(0));
-        Assert.Empty(store.GetRecentRequests(-1));
-    }
 
     // ---------------------------------------------------------------------------
     // GetConfigSnapshot — snapshot timestamp
@@ -1205,234 +948,6 @@ public sealed class StubInspectionServiceTests
     // ---------------------------------------------------------------------------
 
     [Fact]
-    public void GetScenarioStates_ReturnsConfiguredScenarios_WithInitialStateByDefault()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/checkout"] = new()
-                {
-                    Post = new OperationDefinition
-                    {
-                        Responses = new Dictionary<string, ResponseDefinition>(StringComparer.Ordinal)
-                        {
-                            ["409"] = new()
-                            {
-                                Scenario = new ScenarioDefinition
-                                {
-                                    Name = "checkout-flow",
-                                    State = "initial",
-                                    Next = "confirmed"
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        };
-
-        var scenario = Assert.Single(CreateService(document).GetScenarioStates());
-
-        Assert.Equal("checkout-flow", scenario.Name);
-        Assert.Equal("initial", scenario.CurrentState);
-        Assert.Null(scenario.LastUpdatedTimestamp);
-    }
-
-    [Fact]
-    public void GetScenarioStates_ReflectsAdvancedScenarioStateAndTimestamp()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/checkout"] = new()
-                {
-                    Post = new OperationDefinition
-                    {
-                        Responses = new Dictionary<string, ResponseDefinition>(StringComparer.Ordinal)
-                        {
-                            ["409"] = new()
-                            {
-                                Scenario = new ScenarioDefinition
-                                {
-                                    Name = "checkout-flow",
-                                    State = "initial",
-                                    Next = "confirmed"
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        };
-        var loader = new TestStubDefinitionLoader(document);
-        var scenarioService = new ScenarioService();
-        var state = new StubDefinitionState(loader, scenarioService, NullLogger<StubDefinitionState>.Instance);
-        var settings = Options.Create(new StubSettings());
-        var stubService = CreateStubService(state, scenarioService, CreateMatcherService(), new NoOpSemanticMatcherService());
-        var service = new StubInspectionService(
-            state,
-            loader,
-            settings,
-            stubService,
-            new StubInspectionRuntimeStore(),
-            new StubInspectionScenarioCoordinator(state, scenarioService));
-
-        scenarioService.Advance(new ScenarioDefinition
-        {
-            Name = "checkout-flow",
-            State = "initial",
-            Next = "confirmed"
-        });
-
-        var scenario = Assert.Single(service.GetScenarioStates());
-
-        Assert.Equal("confirmed", scenario.CurrentState);
-        Assert.NotNull(scenario.LastUpdatedTimestamp);
-    }
-
-    [Fact]
-    public void ResetScenarioState_ReturnsFalse_WhenScenarioDoesNotExist()
-    {
-        var service = CreateService(EmptyDocument());
-
-        var reset = service.ResetScenarioState("missing");
-
-        Assert.False(reset);
-    }
-
-    [Fact]
-    public void ResetScenarioStates_ResetsAllConfiguredScenariosWithTimestamp()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/checkout"] = new()
-                {
-                    Post = new OperationDefinition
-                    {
-                        Responses = new Dictionary<string, ResponseDefinition>(StringComparer.Ordinal)
-                        {
-                            ["409"] = new()
-                            {
-                                Scenario = new ScenarioDefinition
-                                {
-                                    Name = "checkout-flow",
-                                    State = "initial",
-                                    Next = "confirmed"
-                                }
-                            }
-                        },
-                        Matches =
-                        [
-                            new QueryMatchDefinition
-                            {
-                                Response = new QueryMatchResponseDefinition
-                                {
-                                    StatusCode = 200,
-                                    Scenario = new ScenarioDefinition
-                                    {
-                                        Name = "payment-flow",
-                                        State = "initial",
-                                        Next = "authorized"
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            }
-        };
-        var loader = new TestStubDefinitionLoader(document);
-        var scenarioService = new ScenarioService();
-        var state = new StubDefinitionState(loader, scenarioService, NullLogger<StubDefinitionState>.Instance);
-        var settings = Options.Create(new StubSettings());
-        var stubService = CreateStubService(state, scenarioService, CreateMatcherService(), new NoOpSemanticMatcherService());
-        var service = new StubInspectionService(
-            state,
-            loader,
-            settings,
-            stubService,
-            new StubInspectionRuntimeStore(),
-            new StubInspectionScenarioCoordinator(state, scenarioService));
-
-        scenarioService.Advance(new ScenarioDefinition { Name = "checkout-flow", State = "initial", Next = "confirmed" });
-        scenarioService.Advance(new ScenarioDefinition { Name = "payment-flow", State = "initial", Next = "authorized" });
-
-        service.ResetScenarioStates();
-
-        var scenarios = service.GetScenarioStates();
-
-        Assert.Equal(2, scenarios.Count);
-        Assert.All(scenarios, scenario =>
-        {
-            Assert.Equal("initial", scenario.CurrentState);
-            Assert.NotNull(scenario.LastUpdatedTimestamp);
-        });
-    }
-
-    [Fact]
-    public void StubInspectionScenarioCoordinator_ResetScenarioState_ReturnsFalse_WhenScenarioDoesNotExist()
-    {
-        var loader = new TestStubDefinitionLoader(EmptyDocument());
-        var scenarioService = new ScenarioService();
-        var state = new StubDefinitionState(loader, scenarioService, NullLogger<StubDefinitionState>.Instance);
-        var coordinator = new StubInspectionScenarioCoordinator(state, scenarioService);
-
-        var reset = coordinator.ResetScenarioState("missing");
-
-        Assert.False(reset);
-    }
-
-    [Fact]
-    public void StubInspectionScenarioCoordinator_GetScenarioStates_ReflectsAdvancedScenarioState()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/checkout"] = new()
-                {
-                    Post = new OperationDefinition
-                    {
-                        Responses = new Dictionary<string, ResponseDefinition>(StringComparer.Ordinal)
-                        {
-                            ["409"] = new()
-                            {
-                                Scenario = new ScenarioDefinition
-                                {
-                                    Name = "checkout-flow",
-                                    State = "initial",
-                                    Next = "confirmed",
-                                },
-                            },
-                        },
-                    },
-                },
-            },
-        };
-        var loader = new TestStubDefinitionLoader(document);
-        var scenarioService = new ScenarioService();
-        var state = new StubDefinitionState(loader, scenarioService, NullLogger<StubDefinitionState>.Instance);
-        var coordinator = new StubInspectionScenarioCoordinator(state, scenarioService);
-
-        scenarioService.Advance(new ScenarioDefinition
-        {
-            Name = "checkout-flow",
-            State = "initial",
-            Next = "confirmed",
-        });
-
-        var scenario = Assert.Single(coordinator.GetScenarioStates());
-
-        Assert.Equal("checkout-flow", scenario.Name);
-        Assert.Equal("confirmed", scenario.CurrentState);
-        Assert.NotNull(scenario.LastUpdatedTimestamp);
-    }
-
-    [Fact]
     public void GetLastMatchExplanation_ReturnsNull_WhenNothingHasBeenRecorded()
     {
         var service = CreateService(EmptyDocument());
@@ -1921,14 +1436,4 @@ public sealed class StubInspectionServiceTests
         }
     }
 
-    [Fact]
-    public void StubInspectionRuntimeStore_GetLastMatchExplanation_ReturnsRecordedInstance()
-    {
-        var store = new StubInspectionRuntimeStore();
-        var explanation = CreateRecordedExplanation(matched: true, matchResult: "Matched", routeId: "listUsers");
-
-        store.RecordLastMatchExplanation(explanation);
-
-        Assert.Same(explanation, store.GetLastMatchExplanation());
-    }
 }
