+++ tests/SemanticStub.Api.Tests/Unit/Resolution/StubServiceTests.cs
@@ -584,74 +584,6 @@ public sealed class StubServiceTests
     }
 
     [Fact]
-    public async Task DispatchAsync_MatchesMultiValueQueryParameter()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/search"] = new()
-                {
-                    Get = new OperationDefinition
-                    {
-                        Matches =
-                        [
-                            new QueryMatchDefinition
-                            {
-                                Query = new Dictionary<string, object?>(StringComparer.Ordinal)
-                                {
-                                    ["tag"] = new List<object?> { "alpha", "beta" }
-                                },
-                                Response = new QueryMatchResponseDefinition
-                                {
-                                    StatusCode = 200,
-                                    Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                    {
-                                        ["application/json"] = new()
-                                        {
-                                            Example = new Dictionary<object, object>
-                                            {
-                                                ["result"] = "ordered"
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        ],
-                        Responses = new Dictionary<string, ResponseDefinition>(StringComparer.Ordinal)
-                        {
-                            ["200"] = new()
-                            {
-                                Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                {
-                                    ["application/json"] = new()
-                                    {
-                                        Example = new Dictionary<object, object>
-                                        {
-                                            ["result"] = "fallback"
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        };
-
-        var service = CreateService(document);
-        var query = new Dictionary<string, StringValues>(StringComparer.Ordinal)
-        {
-            ["tag"] = new StringValues(["alpha", "beta"])
-        };
-
-        var (matched, response) = await DispatchAsync(service, HttpMethods.Get, "/search", query);
-        var matchedResponse = AssertMatchedResponse(matched, response);
-
-        Assert.Equal("{\"result\":\"ordered\"}", matchedResponse.Body);
-    }
-
-    [Fact]
     public async Task DispatchAsync_UsesResponseFileContentWhenConfigured()
     {
         var document = new StubDocument
@@ -1079,216 +1011,6 @@ public sealed class StubServiceTests
         var matchedResponse = AssertMatchedResponse(matched, response);
 
         Assert.Equal("{\"result\":\"deleted\"}", matchedResponse.Body);
-    }
-
-    [Fact]
-    public async Task DispatchAsync_PrefersMoreSpecificQueryMatch()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/users"] = new()
-                {
-                    Get = new OperationDefinition
-                    {
-                        Matches =
-                        [
-                            new QueryMatchDefinition
-                            {
-                                Query = new Dictionary<string, object?>(StringComparer.Ordinal)
-                                {
-                                    ["role"] = "admin"
-                                },
-                                Response = new QueryMatchResponseDefinition
-                                {
-                                    StatusCode = 200,
-                                    Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                    {
-                                        ["application/json"] = new()
-                                        {
-                                            Example = new Dictionary<object, object>
-                                            {
-                                                ["message"] = "admin"
-                                            }
-                                        }
-                                    }
-                                }
-                            },
-                            new QueryMatchDefinition
-                            {
-                                Query = new Dictionary<string, object?>(StringComparer.Ordinal)
-                                {
-                                    ["role"] = "admin",
-                                    ["view"] = "summary"
-                                },
-                                Response = new QueryMatchResponseDefinition
-                                {
-                                    StatusCode = 200,
-                                    Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                    {
-                                        ["application/json"] = new()
-                                        {
-                                            Example = new Dictionary<object, object>
-                                            {
-                                                ["message"] = "admin-summary"
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            }
-        };
-
-        var service = CreateService(document);
-        var query = new Dictionary<string, string>(StringComparer.Ordinal)
-        {
-            ["role"] = "admin",
-            ["view"] = "summary"
-        };
-
-        var (matched, response) = await DispatchAsync(service, HttpMethods.Get, "/users", query);
-        var matchedResponse = AssertMatchedResponse(matched, response);
-
-        Assert.Equal("{\"message\":\"admin-summary\"}", matchedResponse.Body);
-    }
-
-    [Fact]
-    public async Task DispatchAsync_FallsBackToDefaultResponseWhenNoQueryMatchExists()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/users"] = new()
-                {
-                    Get = new OperationDefinition
-                    {
-                        Matches =
-                        [
-                            new QueryMatchDefinition
-                            {
-                                Query = new Dictionary<string, object?>(StringComparer.Ordinal)
-                                {
-                                    ["role"] = "admin"
-                                },
-                                Response = new QueryMatchResponseDefinition
-                                {
-                                    StatusCode = 200,
-                                    Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                    {
-                                        ["application/json"] = new()
-                                        {
-                                            Example = new Dictionary<object, object>
-                                            {
-                                                ["message"] = "admin"
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        ],
-                        Responses = new Dictionary<string, ResponseDefinition>(StringComparer.Ordinal)
-                        {
-                            ["200"] = new()
-                            {
-                                Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                {
-                                    ["application/json"] = new()
-                                    {
-                                        Example = new Dictionary<object, object>
-                                        {
-                                            ["message"] = "default"
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        };
-
-        var service = CreateService(document);
-        var query = new Dictionary<string, string>(StringComparer.Ordinal)
-        {
-            ["role"] = "guest"
-        };
-
-        var (matched, response) = await DispatchAsync(service, HttpMethods.Get, "/users", query);
-        var matchedResponse = AssertMatchedResponse(matched, response);
-
-        Assert.Equal("{\"message\":\"default\"}", matchedResponse.Body);
-    }
-
-    [Fact]
-    public async Task DispatchAsync_ReturnsResponseNotConfigured_WhenMatchedQueryResponseIsInvalid()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/users"] = new()
-                {
-                    Get = new OperationDefinition
-                    {
-                        Matches =
-                        [
-                            new QueryMatchDefinition
-                            {
-                                Query = new Dictionary<string, object?>(StringComparer.Ordinal)
-                                {
-                                    ["role"] = "admin"
-                                },
-                                Response = new QueryMatchResponseDefinition
-                                {
-                                    StatusCode = 0,
-                                    Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                    {
-                                        ["application/json"] = new()
-                                        {
-                                            Example = new Dictionary<object, object>
-                                            {
-                                                ["message"] = "broken"
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        ],
-                        Responses = new Dictionary<string, ResponseDefinition>(StringComparer.Ordinal)
-                        {
-                            ["200"] = new()
-                            {
-                                Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                {
-                                    ["application/json"] = new()
-                                    {
-                                        Example = new Dictionary<object, object>
-                                        {
-                                            ["message"] = "default"
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        };
-
-        var service = CreateService(document);
-        var query = new Dictionary<string, string>(StringComparer.Ordinal)
-        {
-            ["role"] = "admin"
-        };
-
-        var (matched, _) = await DispatchAsync(service, HttpMethods.Get, "/users", query);
-
-        Assert.Equal(StubMatchResult.ResponseNotConfigured, matched);
     }
 
     [Fact]
@@ -2173,159 +1895,4 @@ public sealed class StubServiceTests
         Assert.Equal("{\"message\":\"default\"}", matchedResponse.Body);
     }
 
-    [Fact]
-    public async Task DispatchAsync_AdvancesScenarioStateAcrossRequests()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/checkout"] = new()
-                {
-                    Post = new OperationDefinition
-                    {
-                        Responses = new Dictionary<string, ResponseDefinition>(StringComparer.Ordinal)
-                        {
-                            ["409"] = new()
-                            {
-                                Scenario = new ScenarioDefinition
-                                {
-                                    Name = "checkout-flow",
-                                    State = "initial",
-                                    Next = "confirmed"
-                                },
-                                Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                {
-                                    ["application/json"] = new()
-                                    {
-                                        Example = new Dictionary<object, object>
-                                        {
-                                            ["result"] = "pending"
-                                        }
-                                    }
-                                }
-                            },
-                            ["200"] = new()
-                            {
-                                Scenario = new ScenarioDefinition
-                                {
-                                    Name = "checkout-flow",
-                                    State = "confirmed"
-                                },
-                                Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                {
-                                    ["application/json"] = new()
-                                    {
-                                        Example = new Dictionary<object, object>
-                                        {
-                                            ["result"] = "complete"
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        };
-
-        var service = CreateService(document);
-
-        var (firstMatch, firstResponse) = await DispatchAsync(service, HttpMethods.Post, "/checkout");
-        var firstMatchedResponse = AssertMatchedResponse(firstMatch, firstResponse);
-        var (secondMatch, secondResponse) = await DispatchAsync(service, HttpMethods.Post, "/checkout");
-        var secondMatchedResponse = AssertMatchedResponse(secondMatch, secondResponse);
-
-        Assert.Equal(409, firstMatchedResponse.StatusCode);
-        Assert.Equal("{\"result\":\"pending\"}", firstMatchedResponse.Body);
-        Assert.Equal(200, secondMatchedResponse.StatusCode);
-        Assert.Equal("{\"result\":\"complete\"}", secondMatchedResponse.Body);
-    }
-
-    [Fact]
-    public async Task DispatchAsync_FiltersMatchedConditionalResponsesByScenarioState()
-    {
-        var document = new StubDocument
-        {
-            Paths = new Dictionary<string, PathItemDefinition>(StringComparer.Ordinal)
-            {
-                ["/checkout"] = new()
-                {
-                    Post = new OperationDefinition
-                    {
-                        Matches =
-                        [
-                            new QueryMatchDefinition
-                            {
-                                Query = new Dictionary<string, object?>(StringComparer.Ordinal)
-                                {
-                                    ["step"] = "1"
-                                },
-                                Response = new QueryMatchResponseDefinition
-                                {
-                                    StatusCode = 409,
-                                    Scenario = new ScenarioDefinition
-                                    {
-                                        Name = "checkout-flow",
-                                        State = "initial",
-                                        Next = "confirmed"
-                                    },
-                                    Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                    {
-                                        ["application/json"] = new()
-                                        {
-                                            Example = new Dictionary<object, object>
-                                            {
-                                                ["result"] = "pending"
-                                            }
-                                        }
-                                    }
-                                }
-                            },
-                            new QueryMatchDefinition
-                            {
-                                Query = new Dictionary<string, object?>(StringComparer.Ordinal)
-                                {
-                                    ["step"] = "1"
-                                },
-                                Response = new QueryMatchResponseDefinition
-                                {
-                                    StatusCode = 200,
-                                    Scenario = new ScenarioDefinition
-                                    {
-                                        Name = "checkout-flow",
-                                        State = "confirmed"
-                                    },
-                                    Content = new Dictionary<string, MediaTypeDefinition>(StringComparer.Ordinal)
-                                    {
-                                        ["application/json"] = new()
-                                        {
-                                            Example = new Dictionary<object, object>
-                                            {
-                                                ["result"] = "complete"
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            }
-        };
-
-        var service = CreateService(document);
-        var query = new Dictionary<string, string>(StringComparer.Ordinal)
-        {
-            ["step"] = "1"
-        };
-
-        var (firstMatch, firstResponse) = await DispatchAsync(service, HttpMethods.Post, "/checkout", query);
-        var firstMatchedResponse = AssertMatchedResponse(firstMatch, firstResponse);
-        var (secondMatch, secondResponse) = await DispatchAsync(service, HttpMethods.Post, "/checkout", query);
-        var secondMatchedResponse = AssertMatchedResponse(secondMatch, secondResponse);
-
-        Assert.Equal(409, firstMatchedResponse.StatusCode);
-        Assert.Equal(200, secondMatchedResponse.StatusCode);
-    }
 }
